--- scene.cpp
@@ -1,5 +1,0 @@
-//
-// Created by s150818 on 14-2-2019.
-//
-
-#include "scene.h"
